Add to check group permission
- Fix data types

+++ src/main/java/com/page/server/service/UserGroupService.java
@@ -9,7 +9,9 @@ import com.page.server.repository.UserGroupRepository;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
+import org.springframework.util.CollectionUtils;
 
+import javax.transaction.Transactional;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
@@ -22,13 +24,13 @@ public class UserGroupService {
     private final UserGroupRepository userGroupRepository;
     private final PermissionService permissionService;
 
-    public List<UserGroup> getAllUserGroupList() {
-        return userGroupRepository.findAll();
-    }
+//    public List<UserGroup> getAllUserGroupList() {
+//        return userGroupRepository.findAll();
+//    }
 
     public List<UserGroup> getUserGroupListByUser(User user) {
         if (user.isAdmin()) {
-            return getAllUserGroupList();
+            return userGroupRepository.findAll();
         }
 
         List<PermissionDao.No> permissionList = permissionService.getPermissionListByUserNo(user.userNo);
@@ -38,28 +40,60 @@ public class UserGroupService {
         );
     }
 
-    public UserGroup getUserGroup(Long groupNo) {
-        return userGroupRepository.findById(groupNo)
-                .orElseThrow(() -> new IllegalArgumentException("Can't Find Group."));
-    }
+    public UserGroup addUserGroup(User user, UserGroup userGroup) {
+        Permission writePermission = permissionService.createIfNotExist(
+                Permission.builder()
+                        .userNo(user.userNo)
+                        .accessRight(AccessRight.WRITE)
+                        .build()
+        );
 
-    public UserGroup updateUserGroup(UserGroup userGroup) {
-        userGroupRepository.findById(userGroup.groupNo)
-                .orElseThrow(() -> new IllegalArgumentException("Can't Find Group."));
+        if (!user.isAdmin() && CollectionUtils.isEmpty(userGroup.permissions)) {
+            userGroup.permissions = new ArrayList<>();
+        }
+
+        userGroup.permissions.add(writePermission);
 
         return userGroupRepository.save(userGroup);
     }
 
-    public Boolean deleteUserGroup(Long groupNo) {
-        userGroupRepository.findById(groupNo)
+    public UserGroup getUserGroup(User user, Long groupNo) {
+        UserGroup userGroup = userGroupRepository.findById(groupNo)
                 .orElseThrow(() -> new IllegalArgumentException("Can't Find Group."));
+
+        if (!user.isAdmin() && !userGroup.isReadable(user.userNo, null)) {
+            throw new RuntimeException("You don't have permission.");
+        }
+
+        return userGroup;
+    }
+
+    @Transactional
+    public UserGroup updateUserGroup(User user, UserGroup userGroup) {
+        UserGroup oldUserGroup = userGroupRepository.findById(userGroup.groupNo)
+                .orElseThrow(() -> new IllegalArgumentException("Can't Find Group."));
+
+        if (!user.isAdmin() && !oldUserGroup.isWritable(user.userNo, null)) {
+            throw new RuntimeException("You don't have permission.");
+        }
+
+        return userGroupRepository.save(userGroup);
+    }
+
+    public Boolean deleteUserGroup(User user, Long groupNo) {
+        UserGroup oldUserGroup = userGroupRepository.findById(groupNo)
+                .orElseThrow(() -> new IllegalArgumentException("Can't Find Group."));
+
+        if (!user.isAdmin() && oldUserGroup.isWritable(user.userNo, null)) {
+            throw new RuntimeException("You don't have permission.");
+        }
 
         userGroupRepository.deleteById(groupNo);
 
         return userGroupRepository.findById(groupNo).isPresent();
     }
 
-    public void addGroupPermissionIfNotExist(Long userNo, List<Long> groupNos) {
+    public void addGroupListToReadIfNotExist(Long userNo, List<Long> groupNos) {
         List<UserGroup> userGroupList = new ArrayList<>();
 
         Permission permission = permissionService.createIfNotExist(
